+++ jdmp-weka/trunk/src/main/java/org/jdmp/weka/wrappers/SampleToInstanceWrapper.java
@@ -19,7 +19,7 @@ public class SampleToInstanceWrapper extends Instance {
 		if (input != null) {
 			for (int i = 0; i < input.getColumnCount(); i++) {
 				if (discrete) {
-					setValue(i, input.getInt(0, i));
+					setValue(i, (int) input.getDouble(0, i));
 				} else {
 					setValue(i, input.getDouble(0, i));
 				}
